+++ RustyDolphin/src/Networks/Packets/Types/TCPOptions/TCPSACK.h
@@ -6,8 +6,8 @@ class TCPSACK : public TCPOption {
 public:
 	int m_len;
 	int m_edges;
-	long long* m_Redges;
-	long long* m_Ledges;
+	unsigned int* m_Redges;
+	unsigned int* m_Ledges;
 
 	TCPSACK(pcap_pkthdr* header, const u_char* pkt_data, unsigned int* pos);
